hogwarts-artifacts-online/src/main/java/io/github/nkacc/mvnp/hogwartsartifactsonline/artifact/ArtifactService.java
package io.github.nkacc.mvnp.hogwartsartifactsonline.artifact;

import jakarta.transaction.Transactional;
import org.springframework.stereotype.Service;

@Service
@Transactional
public class ArtifactService {

    private final ArtifactRepository artifactRepository;

    public ArtifactService(ArtifactRepository artifactRepository) {
        this.artifactRepository = artifactRepository;
    }

    public Artifact findById(String artifactId){
        return artifactRepository.findById(artifactId).get();
    }

}
